add feature Dynamic Menus
