+++ src/components/todo/ErrorComponent.jsx
@@ -1,0 +1,8 @@
+import React, {Component} from 'react'
+import AuthenticationService from './AuthenticationService.js'
+function ErrorComponent(){
+    
+    return <div>An error ocurred. Pls contact admin</div>
+    
+}
+export default ErrorComponent
