+++ src/components/todo/FooterComponent.jsx
@@ -1,0 +1,14 @@
+import React, {Component} from 'react'
+
+import AuthenticationService from './AuthenticationService.js'
+class FooterComponent extends Component{
+    render(){
+        return(
+            <footer className="footer">
+                <span className="text-muted">All Rights Reserved 2021 @HongSDE</span>
+            
+            </footer>
+        )
+    }
+}    
+export default FooterComponent
